rename index.ts to site.ts
