+++ packages/docusaurus/src/commands/deploy.ts
@@ -11,7 +11,7 @@ import os from 'os';
 import logger from '@docusaurus/logger';
 import shell from 'shelljs';
 import {hasSSHProtocol, buildSshUrl, buildHttpsUrl} from '@docusaurus/utils';
-import {loadContext, type LoadContextParams} from '../server';
+import {loadContext, type LoadContextParams} from '../server/site';
 import {build} from './build';
 
 export type DeployCLIOptions = Pick<
