+++ packages/docusaurus/src/commands/serve.ts
@@ -15,7 +15,7 @@ import openBrowser from 'react-dev-utils/openBrowser';
 import {loadSiteConfig} from '../server/config';
 import {build} from './build';
 import {getHostPort, type HostPortOptions} from '../server/getHostPort';
-import type {LoadContextParams} from '../server';
+import type {LoadContextParams} from '../server/site';
 
 export type ServeCLIOptions = HostPortOptions &
   Pick<LoadContextParams, 'config'> & {
